split out SFC's
fixed filter and made it use simpler formcontrol
