+++ src/client/app/heroes/hero-detail/hero-detail.component.scss
