+++ src/client/app/heroes/hero-list/hero-list.component.html
@@ -1,0 +1,34 @@
+<div>
+  <div>
+    <div class="button-group">
+      <button type="button" (click)="getHeroes()">Refresh</button>
+      <button type="button" (click)="enableAddMode()" *ngIf="!addingHero && !selectedHero">Add</button>
+    </div>
+    <div>
+      <p>Filter the heroes</p>
+      <input [formControl]="filter" (input)="setFilter(filter.value)" />
+    </div>
+    <div *ngIf="filteredHeroes$ | async as heroes">
+      <div *ngIf="loading$ | async;else heroList">Loading</div>
+      <ng-template #heroList>
+        <ul class="heroes">
+          <li *ngFor="let hero of heroes" class="hero-container" [class.selected]="hero === selectedHero">
+            <div class="hero-element">
+              <div class="badge">{{hero.id}}</div>
+              <div class="hero-text" (click)="onSelect(hero)">
+                <div class="name">{{hero.name}}</div>
+                <div class="saying">{{hero.saying}}</div>
+              </div>
+            </div>
+            <button class="delete-button" (click)="deleteHero(hero)">Delete</button>
+          </li>
+        </ul>
+      </ng-template>
+    </div>
+  </div>
+  <div>
+    <ng-template #elseTemplate>Loading ...</ng-template>
+    <app-hero-detail *ngIf="selectedHero || addingHero" [hero]="selectedHero" (unselect)="unselect()" (add)="add($event)" (update)="update($event)">
+    </app-hero-detail>
+  </div>
+</div>
