+++ src/client/app/heroes/hero-list/hero-list.component.scss
