+++ src/client/app/heroes/hero-list/hero-list.component.ts
@@ -1,64 +1,23 @@
 import { Component, OnInit, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { FormControl } from '@angular/forms';
 
-import { AppSelectors } from '../store/app-config';
+import { AppSelectors } from '../../store/app-config';
 import {
   EntityDispatchers,
   EntityDispatcher,
   EntitySelectors,
   EntitySelector
-} from '../../ngrx-data';
+} from '../../../ngrx-data';
 
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { debounceTime, distinctUntilChanged, skip, takeUntil } from 'rxjs/operators';
 
-import { Hero } from '../core';
+import { Hero } from '../../core';
 
 @Component({
   selector: 'app-hero-list',
-  template: `
-    <div>
-      <form [formGroup]="form">
-      <div class="button-group">
-          <button type="button" (click)="getHeroes()">Refresh</button>
-          <button type="button" (click)="enableAddMode()" *ngIf="!addingHero && !selectedHero">Add</button>
-        </div>
-        <div>
-          <p>Filter the heroes</p>
-          <input formControlName="filter" (input)="setFilter(form)"/>
-        </div>
-        <div *ngIf="filteredHeroes$ | async as heroes">
-          <div *ngIf="loading$ | async;else heroList">Loading</div>
-          <ng-template #heroList>
-            <ul class="heroes">
-              <li *ngFor="let hero of heroes"
-                class="hero-container"
-                [class.selected]="hero === selectedHero">
-                <div class="hero-element">
-                  <div class="badge">{{hero.id}}</div>
-                  <div class="hero-text" (click)="onSelect(hero)">
-                    <div class="name">{{hero.name}}</div>
-                    <div class="saying">{{hero.saying}}</div>
-                  </div>
-                </div>
-                <button class="delete-button" (click)="deleteHero(hero)">Delete</button>
-              </li>
-            </ul>
-          </ng-template>
-        </div>
-      </form>
-
-      <ng-template #elseTemplate>Loading ...</ng-template>
-      <app-hero-detail
-        *ngIf="selectedHero || addingHero"
-        [hero]="selectedHero"
-        (unselect)="unselect()"
-        (add)="add($event)"
-        (update)="update($event)">
-      </app-hero-detail>
-    </div>
-  `,
+  templateUrl: './hero-list.component.html',
   styleUrls: ['./hero-list.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
@@ -70,16 +29,13 @@ export class HeroListComponent implements OnDestroy, OnInit {
   loading$: Observable<boolean>;
   filter$: Observable<string>;
   dataSource$ = this.appSelectors.dataSource$();
-  form = this.fb.group({
-    filter: ['']
-  });
+  filter: FormControl = new FormControl();
 
   private onDestroy = new Subject();
   private heroDispatcher: EntityDispatcher<Hero>;
   private heroSelector: EntitySelector<Hero>;
 
   constructor(
-    private fb: FormBuilder,
     entityDispatchers: EntityDispatchers,
     entitySelectors: EntitySelectors,
     private appSelectors: AppSelectors
@@ -95,22 +51,20 @@ export class HeroListComponent implements OnDestroy, OnInit {
 
     this.dataSource$
       .pipe(takeUntil(this.onDestroy), distinctUntilChanged())
-      .subscribe((val: string) => this.getHeroes());
+      .subscribe(value => this.getHeroes());
 
-    this.filter$
-      .pipe(takeUntil(this.onDestroy), debounceTime(500), distinctUntilChanged())
-      .subscribe((val: string) => this.filterHeroes());
+    this.filter$.pipe(takeUntil(this.onDestroy), distinctUntilChanged()).subscribe(value => {
+      this.filter.setValue(value);
+      this.filterHeroes();
+    });
   }
 
   ngOnDestroy() {
     this.onDestroy.next(true);
   }
 
-  setFilter(form: FormGroup) {
-    const { value, valid, touched } = form;
-    if (valid) {
-      this.heroDispatcher.setFilter(value.filter);
-    }
+  setFilter(value: string) {
+    this.heroDispatcher.setFilter(value);
     this.clear();
   }
