+++ src/client/app/heroes/heroes-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { HeroListComponent } from './hero-list.component';
+import { HeroListComponent } from './hero-list/hero-list.component';
 
 const routes: Routes = [{ path: '', pathMatch: 'full', component: HeroListComponent }];
