+++ src/client/app/villains/villain-detail/villain-detail.component.html
@@ -1,0 +1,20 @@
+<div class="editarea">
+  <form [formGroup]="form">
+    <div class="editfields">
+      <div [hidden]="addMode">
+        <label>id: </label>
+        <input formControlName="id" #id readonly/>
+      </div>
+      <div>
+        <label>name: </label>
+        <input formControlName="name" placeholder="name" #name />
+      </div>
+      <div>
+        <label>saying: </label>
+        <input formControlName="saying" placeholder="saying" (keyup.enter)="saveVillain(form)" />
+      </div>
+    </div>
+    <button type="button" (click)="clear()">Cancel</button>
+    <button type="button" (click)="saveVillain(form)">Save</button>
+  </form>
+</div>
