+++ src/client/app/villains/villain-detail/villain-detail.component.ts
@@ -8,37 +8,18 @@ import {
   OnInit,
   Output,
   ViewChild,
-  SimpleChanges
+  SimpleChanges,
+  ChangeDetectionStrategy
 } from '@angular/core';
 
-import { Villain } from '../core';
+import { Villain } from '../../core';
 import { FormBuilder, Validators, FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-villain-detail',
-  template: `
-    <div class="editarea">
-      <form [formGroup]="form">
-        <div class="editfields">
-          <div [hidden]="addMode">
-            <label>id: </label>
-            <input formControlName="id" #id readonly/>
-          </div>
-          <div>
-            <label>name: </label>
-            <input formControlName="name" placeholder="name" #name />
-          </div>
-          <div>
-            <label>saying: </label>
-            <input formControlName="saying" placeholder="saying" (keyup.enter)="saveVillain(form)"/>
-          </div>
-        </div>
-        <button type="button" (click)="clear()">Cancel</button>
-        <button type="button" (click)="saveVillain(form)">Save</button>
-      </form>
-    </div>
-    `,
-  styleUrls: ['./villain-detail.component.scss']
+  templateUrl: './villain-detail.component.html',
+  styleUrls: ['./villain-detail.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class VillainDetailComponent implements AfterViewInit, OnChanges, OnInit {
   @Input() villain: Villain;
