+++ src/client/app/villains/villain-list/villain-list.component.html
@@ -1,0 +1,35 @@
+<div>
+  <div>
+    <div class="button-group">
+      <button type="button" (click)="getVillains()">Refresh</button>
+      <button type="button" (click)="enableAddMode()" *ngIf="!addingVillain && !selectedVillain">Add</button>
+    </div>
+    <div>
+      <p>Filter the villains</p>
+      <input [formControl]="filter" (input)="setFilter(filter.value)" />
+    </div>
+    <div *ngIf="filteredVillains$ | async as villains">
+      <div *ngIf="loading$ | async;else villainList">Loading</div>
+      <ng-template #villainList>
+        <ul class="villains">
+          <li *ngFor="let villain of villains" class="villain-container" [class.selected]="villain === selectedVillain">
+            <div class="villain-element">
+              <div class="badge">{{villain.id}}</div>
+              <div class="villain-text" (click)="onSelect(villain)">
+                <div class="name">{{villain.name}}</div>
+                <div class="saying">{{villain.saying}}</div>
+              </div>
+            </div>
+            <button class="delete-button" (click)="deleteVillain(villain)">Delete</button>
+          </li>
+        </ul>
+      </ng-template>
+    </div>
+  </div>
+  <div>
+    <ng-template #elseTemplate>Loading ...</ng-template>
+    <app-villain-detail *ngIf="selectedVillain || addingVillain" [villain]="selectedVillain" (unselect)="unselect()" (add)="add($event)"
+      (update)="update($event)">
+    </app-villain-detail>
+  </div>
+</div>
