+++ src/client/app/villains/villains.module.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { VillainsRoutingModule } from './villains-routing.module';
-import { VillainListComponent } from './villain-list.component';
-import { VillainDetailComponent } from './villain-detail.component';
+import { VillainListComponent } from './villain-list/villain-list.component';
+import { VillainDetailComponent } from './villain-detail/villain-detail.component';
 
 @NgModule({
   imports: [CommonModule, ReactiveFormsModule, VillainsRoutingModule],
